requirements.txt
configparser==6.0.0
prettytable==3.9.0
psycopg2-binary==2.9.7
